Backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
motor==3.3.1
pymongo==4.6.3
pydantic[email]==1.10.12
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
fastapi-mail==1.3.1 #updated version
python-multipart==0.0.9
Pillow==10.2.0
cloudinary==1.38.0
tensorflow-cpu==2.19.0
fastapi-jwt-auth==0.5.0
keras==3.10.0

# uvicorn app.main:app --host 0.0.0.0 --port 8000
# uvicorn app.main:app --host 127.0.0.1 --port 8000

# venv\Scripts\activate.bat 
# http://localhost:8000